requirements.txt
bottle==0.12.9
docker-py==1.6.0
PyYAML==3.11
click==6.2
celery==3.1.19
